+++ src/y2019/day7/Part2.java
@@ -3,20 +3,20 @@ package y2019.day7;
 import utils.Problem;
 import y2019.intcode.Parser;
 
-public class Part2 extends Problem<Integer> {
+public class Part2 extends Problem<Long> {
     private Parser parser;
 
-    public Part2(String fileName, int exp) {
+    public Part2(String fileName, long exp) {
         super(exp);
 
         this.parser = new Parser(fileName);
     }
 
-    public Integer run() {
+    public Long run() {
         var prog = parser.parse();
         var perms = Utils.findPerms(new int[] { 5, 6, 7, 8, 9 });
 
-        var most = 0;
+        var most = 0L;
 
         for (var perm : perms) {
             var circuit = new Circuit(prog, 5);
